+++ lib/main.dart
@@ -1,5 +1,5 @@
 import 'package:customneon/app_bindings.dart';
-import 'package:customneon/homepage/homepage.dart';
+import 'package:customneon/screens/homepage/homepage.dart';
 import 'package:firebase_core/firebase_core.dart';
 import 'package:flutter/material.dart';
 import 'package:get/get.dart';
