--- lib/responsive/dimensions.dart
@@ -1,1 +1,0 @@
-const mobileWidth = 600;
