--- lib/responsive/responsive_layout.dart
@@ -1,24 +1,0 @@
-import 'package:flutter/material.dart';
-
-import 'dimensions.dart';
-
-class ResponsiveLayout extends StatelessWidget {
-  final Widget mobileBody;
-  final Widget desktopBody;
-
-  const ResponsiveLayout(
-      {super.key, required this.mobileBody, required this.desktopBody});
-
-  @override
-  Widget build(BuildContext context) {
-    return LayoutBuilder(
-      builder: (context, constraints) {
-        if (constraints.maxWidth < mobileWidth) {
-          return mobileBody;
-        } else {
-          return desktopBody;
-        }
-      },
-    );
-  }
-}
