+++ lib/screens/create_neon/create_neon.dart
@@ -1,13 +1,13 @@
 import 'package:customneon/controllers/auth_controller.dart';
 import 'package:customneon/controllers/create_neon_controller.dart';
-import 'package:customneon/desktop_view/auth_view/signin_view.dart';
-import 'package:customneon/desktop_view/create_neon/backboard_style_widget.dart';
-import 'package:customneon/desktop_view/create_neon/choose_color_widget.dart';
-import 'package:customneon/desktop_view/create_neon/choose_size_widget.dart';
-import 'package:customneon/desktop_view/create_neon/location.dart';
-import 'package:customneon/desktop_view/create_neon/select_font.dart';
-import 'package:customneon/desktop_view/create_neon/text_align_widget.dart';
-import 'package:customneon/desktop_view/footer/footer_design.dart';
+import 'package:customneon/screens/auth_view/signin_view.dart';
+import 'package:customneon/screens/create_neon/backboard_style_widget.dart';
+import 'package:customneon/screens/create_neon/choose_color_widget.dart';
+import 'package:customneon/screens/create_neon/choose_size_widget.dart';
+import 'package:customneon/screens/create_neon/location.dart';
+import 'package:customneon/screens/create_neon/select_font.dart';
+import 'package:customneon/screens/create_neon/text_align_widget.dart';
+import 'package:customneon/screens/footer/footer_design.dart';
 import 'package:customneon/utills/app_colors.dart';
 import 'package:customneon/utills/app_text_style.dart';
 import 'package:customneon/utills/image_path.dart';
