+++ lib/screens/desktop_view.dart
