+++ lib/screens/footer/faqs.dart
