+++ lib/screens/footer/footer_design.dart
@@ -1,12 +1,12 @@
-import 'package:customneon/desktop_view/footer/about_us.dart';
-import 'package:customneon/desktop_view/footer/contact_us.dart';
-import 'package:customneon/desktop_view/footer/faqs.dart';
-import 'package:customneon/desktop_view/footer/payment_methods.dart';
-import 'package:customneon/desktop_view/footer/privacy_policy.dart';
-import 'package:customneon/desktop_view/footer/refund_policy.dart';
-import 'package:customneon/desktop_view/footer/shipping_policy.dart';
-import 'package:customneon/desktop_view/footer/terms_of_service.dart';
-import 'package:customneon/desktop_view/footer/track_my_order.dart';
+import 'package:customneon/screens/footer/about_us.dart';
+import 'package:customneon/screens/footer/contact_us.dart';
+import 'package:customneon/screens/footer/faqs.dart';
+import 'package:customneon/screens/footer/payment_methods.dart';
+import 'package:customneon/screens/footer/privacy_policy.dart';
+import 'package:customneon/screens/footer/refund_policy.dart';
+import 'package:customneon/screens/footer/shipping_policy.dart';
+import 'package:customneon/screens/footer/terms_of_service.dart';
+import 'package:customneon/screens/footer/track_my_order.dart';
 import 'package:customneon/utills/app_colors.dart';
 import 'package:customneon/utills/app_text_style.dart';
 import 'package:flutter/material.dart';
