+++ lib/screens/footer/payment_methods.dart
