+++ lib/screens/footer/shipping_policy.dart
