+++ lib/screens/homepage/homepage.dart
@@ -1,0 +1,15 @@
+import 'package:customneon/screens/create_neon/create_neon.dart';
+import 'package:flutter/material.dart';
+
+class HomePage extends StatefulWidget {
+  const HomePage({Key? key}) : super(key: key);
+  @override
+  State<HomePage> createState() => _HomePageState();
+}
+
+class _HomePageState extends State<HomePage> {
+  @override
+  Widget build(BuildContext context) {
+    return const Scaffold(body: CreateNeon());
+  }
+}
